stage

+++ src/TD.Web/TD.Web.Common/TD.Web.Common.Contracts/Requests/Users/UpdateUserRequest.cs
@@ -7,14 +7,14 @@ namespace TD.Web.Common.Contracts.Requests.Users
     {
         public string Username { get; set; }
         public string Nickname { get; set; }
-        public int? ProfessionId { get; set; }
+        public long? ProfessionId { get; set; }
         public int? PPID { get; set; }
         public DateTime DateOfBirth { get; set; }
-        public int CityId { get; set; }
+        public long CityId { get; set; }
         public string Address { get; set; }
         public string Mobile { get; set; }
         public string? Mail { get; set; }
-        public int FavoriteStoreId { get; set; }
+        public long FavoriteStoreId { get; set; }
         public string? Comment { get; set; }
         public UserType Type { get; set; }
     }
